+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [2.5.0]
+
+### Added
+
+- Adds support for formatting with `findent` and `fprettify`
+  ([#29](https://github.com/krvajal/vscode-fortran-support/issues/29))
+
 ## [2.4.3]
 
 ### Changed
@@ -22,12 +29,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   ([#257](https://github.com/krvajal/vscode-fortran-support/issues/257))
 - Linting is now operational for `FortranFixedForm`
   ([#258](https://github.com/krvajal/vscode-fortran-support/issues/258))
+- Fixes dummy variable list erroneous syntax highlighting
+  ([#264](https://github.com/krvajal/vscode-fortran-support/issues/264))
 
 ### Changed
 
 - Renamed the Fixed Format Format language from `fortran_fixed-form` to
   `FortranFixedForm`, an alias has been added for backwards compatibility
   ([#259](https://github.com/krvajal/vscode-fortran-support/issues/259))
+
+### Added
+
+- Adds prompts for installing Fortran IntelliSense and fortran-language-server
 
 ### Removed
 
@@ -306,7 +319,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Initial release
 
-[unreleased]: https://github.com/krvajal/vscode-fortran-support/compare/v2.4.3...HEAD
+[unreleased]: https://github.com/krvajal/vscode-fortran-support/compare/v2.5.0...HEAD
+[2.5.0]: https://github.com/krvajal/vscode-fortran-support/compare/v2.4.3...v2.5.0
 [2.4.3]: https://github.com/krvajal/vscode-fortran-support/compare/v2.4.2...v2.4.3
 [2.4.2]: https://github.com/krvajal/vscode-fortran-support/compare/v2.4.1...v2.4.2
 [2.4.1]: https://github.com/krvajal/vscode-fortran-support/compare/v2.4.0...v2.4.1